requirements.txt
asgiref==3.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.4
django-paypal==2.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
Pillow==10.0.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.2.2
whitenoise==6.6.0
dj-database-url==2.1.0dj-database-url==2.1.0
dj-database-url==2.1.0
